+++ src/test/java/api/stepdef/ReqresinStepDef.java
@@ -1,7 +1,8 @@
-package api.stepdef.reqresin;
+package api.stepdef;
 
 import api.service.reqresin.Reqresin;
 import io.cucumber.java.en.And;
+import io.cucumber.java.en.When;
 
 public class ReqresinStepDef {
 
@@ -22,5 +23,19 @@ public class ReqresinStepDef {
     @And("user send GET Users request to reqresin")
     public void getUsers() {
         reqresin.getListUser();
+    }
+
+    @When("user send POST login request to reqresin with body json {string}")
+    public void userSendPostLoginRequestToReqresinWithBodyJson(String jsonPath) {
+        reqresin.postLoginWithParam(jsonPath);
+    }
+
+    @When("user send GET Single request to reqresin with body json {string}")
+    public void userSendGETSingleRequestToReqresinWithBodyJson(String jsonPath) {
+        reqresin.getSingleRequestWithParam(jsonPath);
+    }
+
+    @When("user send POST create request to reqresin")
+    public void postCreate() { reqresin.postCreate();
     }
 }
