+++ src/test/resources/payload/failed-login.json
@@ -1,0 +1,3 @@
+{
+  "email": "eve.holt@reqres.in"
+}
